client: moved all the HTTP helpers into its own concern
This way we have a separation of:

- http_helpers: HTTP helpers and methods for the authentication process.
- registry_client: just the methods dealing with the endpoints themselves.

Moreover, the `get_request` method has been renamed to `perform_request`, since
now it's capable of performing any HTTP method. This has been done so it's
easier to implement features that are still pending to be implemented like
deleting images.
